+++ python/app/models/__init__.py
@@ -1,11 +1,11 @@
-from .Deliverables import Deliverables
-from .ExportShotsThread import ExportShotsThread
-from .Letterbox import Letterbox
-from .LoadShotsThread import LoadShotsThread
-from .NukeProcess import NukeProcess
-from .PreviewOutput import PreviewOutput
-from .SequenceOutput import SequenceOutput
-from .Settings import Settings
-from .Shot import Shot
-from .UserSettings import UserSettings
-from .Version import Version
+from .deliverables import Deliverables
+from .export_shots_thread import ExportShotsThread
+from .letterbox import Letterbox
+from .load_shots_thread import LoadShotsThread
+from .nuke_process import NukeProcess
+from .peview_output import PreviewOutput
+from .sequence_output import SequenceOutput
+from .settings import Settings
+from .shot import Shot
+from .user_settings import UserSettings
+from .version import Version
