+++ python/app/models/export_shots_thread.py
