+++ python/app/models/load_shots_thread.py
